Feat: CSV data type change
+밀리초로 단위변경
+BattleConfig.csv -> InvincibleTime
+SkillTable.csv -> CoolTime, IntervalTime, Duration

+++ Assets/Scripts/InGame/Skill/SkillData.cs
@@ -21,8 +21,8 @@ public class SkillData
     public CALC_DAMAGE_TYPE calcDamageType { get; private set; }    //데미지 계산 방식
     public string skillPrefabPath { get; private set; }
     public int projectileCount { get; private set; }                //투사체 개수
-    public float intervalTime { get; private set; }                 //투사체간 발사 간격
-    public float duration { get; private set; }                     //스킬 지속 시간
+    public int intervalTime { get; private set; }                 //투사체간 발사 간격
+    public int duration { get; private set; }                     //스킬 지속 시간
     public int speed { get; private set; }                          //투사체 속도
     public int splashRange { get; private set; }                    //스플레쉬 범위 (원의 반지름)
     public int projectileSizeMulti { get; private set; }            //투사체 크기 배율
@@ -33,8 +33,8 @@ public class SkillData
     public void SetCoolTime(int coolTime) { this.coolTime = coolTime; }
     public void SetAttackDistance(int attackDistance) { this.attackDistance = attackDistance; }
     public void SetProjectileCount(int projectileCount) { this.projectileCount = projectileCount; }
-    public void SetIntervalTime(float intervalTime) { this.intervalTime = intervalTime; }
-    public void SetDuration(float duration) { this.duration = duration; }
+    public void SetIntervalTime(int intervalTime) { this.intervalTime = intervalTime; }
+    public void SetDuration(int duration) { this.duration = duration; }
     public void SetDamage(int damage) { this.damage = damage; }
     public void SetSpeed(int speed) { this.speed = speed; }
     public void SetSplashRange(int splashRange) { this.splashRange = splashRange; }
